+++ src/rb-radio-control/index.js
@@ -1,7 +1,8 @@
 define([
+    'components/rb-icon',
     './rb-radio-control-directive',
     './rb-radio-control.css'
-], function (rbRadioControlDirective, css) {
+], function (rbIcon, rbRadioControlDirective, css) {
     /**
      * @ngdoc module
      * @name rb-radio-control
@@ -11,7 +12,7 @@ define([
      *
      */
     var rbRadioControl = angular
-        .module('rb-radio-control', [])
+        .module('rb-radio-control', [rbIcon.name])
         .directive('rbRadioControl', rbRadioControlDirective);
 
     return rbRadioControl;